updateHW
