+++ controllers/ctrl.js
@@ -1,12 +1,6 @@
 const contacts = require("../models/contacts");
-const { HttpError } = require("../helpers");
-const Joi = require("joi");
 
-const addSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-});
+const { HttpError } = require("../helpers");
 
 const getAll = async (req, res, next) => {
   try {
@@ -16,6 +10,7 @@ const getAll = async (req, res, next) => {
     next(error);
   }
 };
+
 const getById = async (req, res, next) => {
   try {
     const { id } = req.params;
@@ -28,6 +23,7 @@ const getById = async (req, res, next) => {
     next(error);
   }
 };
+
 const deleteById = async (req, res, next) => {
   try {
     const { id } = req.params;
@@ -42,24 +38,18 @@ const deleteById = async (req, res, next) => {
     next(error);
   }
 };
+
 const addContact = async (req, res, next) => {
   try {
-    const { error } = addSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, "missing required name field");
-    }
     const result = await contacts.addContact(req.body);
     res.status(201).json(result);
   } catch (error) {
     next(error);
   }
 };
+
 const updateContactById = async (req, res, next) => {
   try {
-    const { error } = addSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
     const { id } = req.params;
     const result = await contacts.updateContact(id, req.body);
     if (!result) {
