+++ routes/api/contacts.js
@@ -1,5 +1,9 @@
 const express = require("express");
 const ctrl = require("../../controllers/ctrl");
+const {
+  addContactValidation,
+  updateContactValidation,
+} = require("../../validators/contacts");
 
 const router = express.Router();
 
@@ -7,10 +11,10 @@ router.get("/", ctrl.getAll);
 
 router.get("/:id", ctrl.getById);
 
-router.post("/", ctrl.addContact);
+router.post("/", addContactValidation, ctrl.addContact);
 
 router.delete("/:id", ctrl.deleteById);
 
-router.put("/:id", ctrl.updateContactById);
+router.put("/:id", updateContactValidation, ctrl.updateContactById);
 
 module.exports = router;
